requirements.txt
fastapi>=0.68.0,<0.69.0
uvicorn>=0.15.0,<0.16.0
websockets>=10.0,<11.0
numpy>=1.21.0,<2.0.0
python-dotenv>=0.19.0,<0.20.0
ffmpeg-python==0.2.0
torch
transformers
accelerate
optimum
TTS
